+++ packages/ui-storybook/.storybook/main.mjs
@@ -1,3 +1,4 @@
+import { dirname, join } from "path";
 import { mergeConfig } from 'vite'
 
 export default {
@@ -8,11 +9,12 @@ export default {
     },
   ],
   addons: [
-    '@storybook/addon-a11y',
-    '@storybook/addon-toolbars',
-    '@storybook/addon-viewport',
-    '@storybook/addon-controls',
-    '@saas-ui/storybook-addon',
+    getAbsolutePath("@storybook/addon-a11y"),
+    getAbsolutePath("@storybook/addon-toolbars"),
+    getAbsolutePath("@storybook/addon-viewport"),
+    getAbsolutePath("@storybook/addon-controls"),
+    getAbsolutePath("@saas-ui/storybook-addon"),
+    '@chromatic-com/storybook'
   ],
   staticDirs: ['./static'],
   typescript: {
@@ -42,9 +44,11 @@ export default {
     }
   },
   framework: {
-    name: '@storybook/react-vite',
+    name: getAbsolutePath("@storybook/react-vite"),
   },
-  docs: {
-    autodocs: false,
-  },
+  docs: {},
+}
+
+function getAbsolutePath(value) {
+  return dirname(require.resolve(join(value, "package.json")));
 }