+++ saas-ui/react/src/theme/components/data-grid.ts
@@ -148,10 +148,6 @@ const baseStyle = definePartsStyle({
     padding: 0,
     '&[data-pinned=left]': pinnedLeftStyles,
     '&[data-pinned=right]': pinnedRightStyles,
-    '&[data-is-numeric=true]': {
-      textAlign: 'end',
-      justifyContent: 'end',
-    },
   },
   title: {
     display: 'flex',
@@ -161,8 +157,12 @@ const baseStyle = definePartsStyle({
       transitionProperty: 'opacity',
       transitionDuration: 'normal',
     },
-    '&[aria-sort="none"]:hover, &[aria-sort="none"]:focus svg': {
+    '&[aria-sort="none"]:hover svg, &[aria-sort="none"]:focus-visible svg': {
       opacity: 1,
+    },
+    '[data-is-numeric=true] &': {
+      textAlign: 'end',
+      justifyContent: 'end',
     },
     _focusVisible: {
       outlineColor: 'purple.400',