+++ packages/marketing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # marketing
 
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [32c66c9]
+- Updated dependencies [66411b3]
+- Updated dependencies [ae8b85e]
+- Updated dependencies [ae8b85e]
+- Updated dependencies [32c66c9]
+  - @ui/lib@11.0.1
+  - @saas-ui-pro/react@0.35.6
+
 ## 6.0.0
 
 ### Patch Changes